aave/requirements_aave_bot.txt
requests>=2.28.0
websockets>=11.0.0
python-dotenv>=1.0.0
numpy>=1.21.0
web3>=6.0.0
ccxt>=4.0.0
# Database support
# GPU acceleration (optional)
cupy-cuda12>=12.0.0; sys_platform == 'linux'  # For Linux with CUDA 12
cupy-cuda11>=11.0.0; sys_platform == 'linux' and python_version >= '3.8'  # For Linux with CUDA 11
# CPU acceleration (optional)
numba>=0.56.0
